--- src/Nethermind/Nethermind.Precompiles.Benchmark/ModExp2565Benchmark.cs
@@ -1,35 +1,0 @@
-﻿//  Copyright (c) 2020 Demerzel Solutions Limited
-//  This file is part of the Nethermind library.
-// 
-//  The Nethermind library is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU Lesser General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-// 
-//  The Nethermind library is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-//  GNU Lesser General Public License for more details.
-// 
-//  You should have received a copy of the GNU Lesser General Public License
-//  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
-
-using System.Collections.Generic;
-using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Diagnostics.Windows.Configs;
-using BenchmarkDotNet.Jobs;
-using Nethermind.Evm.Precompiles;
-
-namespace Nethermind.Precompiles.Benchmark
-{
-    [HtmlExporter]
-    // [NativeMemoryProfiler]
-    // [MemoryDiagnoser]
-    // [ShortRunJob(RuntimeMoniker.NetCoreApp50)]
-    [SimpleJob(RuntimeMoniker.NetCoreApp50)]
-    public class ModExp2565Benchmark : PrecompileBenchmarkBase
-    {
-        protected override IEnumerable<IPrecompile> Precompiles => new[] {ModExpPrecompile2565.Instance};
-        protected override string InputsDirectory => "modexp";
-    }
-}
