+++ src/Nethermind/Nethermind.Precompiles.Benchmark/ModExpBenchmark.cs
@@ -16,7 +16,6 @@
 
 using System.Collections.Generic;
 using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Diagnostics.Windows.Configs;
 using BenchmarkDotNet.Jobs;
 using Nethermind.Evm.Precompiles;
